TextBlock::onKey => InnerTextBlock::deleteOutOf
fix backspace and arrow keys in TextBlocks, helps with #121

Oh my god, deleteOutOf is so much prettier, simpler, and cleaner than
doing key checking logic in onKey! And it, y'know, works, which also
technically counts as improvement.

Discovering abstraction barriers and watching it simplify everything is
such a magical feeling.